requirements.txt
pywin32==306
python-dotenv>=1.0.1
openpyxl>=3.1.2
beautifulsoup4>=4.12.3
requests>=2.32.3